use size_t for parse depth

+++ src/aabro.c
@@ -579,13 +579,13 @@ char *abr_parser_to_s(abr_parser *p)
 //
 // the parse methods
 
-typedef abr_tree *abr_p_func(const char *, size_t, int, abr_parser *);
+typedef abr_tree *abr_p_func(const char *, size_t, size_t, abr_parser *);
 //
 abr_tree *abr_do_parse(
-  const char *input, size_t offset, int depth, abr_parser *p);
+  const char *input, size_t offset, size_t depth, abr_parser *p);
 
 abr_tree *abr_p_string(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   char *s = p->string;
   size_t le = p->string_length;
@@ -602,7 +602,7 @@ abr_tree *abr_p_string(
 }
 
 abr_tree *abr_p_regex(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   regmatch_t ms[1];
 
@@ -617,7 +617,7 @@ abr_tree *abr_p_regex(
 }
 
 abr_tree *abr_p_rep(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   short result = 1;
   size_t off = offset;
@@ -650,7 +650,7 @@ abr_tree *abr_p_rep(
 }
 
 abr_tree *abr_p_alt(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   short result = 0;
   size_t length = 0;
@@ -677,7 +677,7 @@ abr_tree *abr_p_alt(
 }
 
 abr_tree *abr_p_seq(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   short result = 1;
   size_t length = 0;
@@ -705,14 +705,14 @@ abr_tree *abr_p_seq(
 }
 
 abr_tree *abr_p_not(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   // not yet implemented
   return NULL;
 }
 
 abr_tree *abr_p_name(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   abr_tree *t = abr_do_parse(input, offset, depth + 1, p->children[0]);
 
@@ -720,21 +720,21 @@ abr_tree *abr_p_name(
 }
 
 abr_tree *abr_p_presence(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   // not yet implemented
   return NULL;
 }
 
 abr_tree *abr_p_absence(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   // not yet implemented
   return NULL;
 }
 
 abr_tree *abr_p_n(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   if (p->children == NULL)
   {
@@ -760,7 +760,7 @@ abr_p_func *abr_p_funcs[] = { // const ?
 };
 
 abr_tree *abr_do_parse(
-  const char *input, size_t offset, int depth, abr_parser *p)
+  const char *input, size_t offset, size_t depth, abr_parser *p)
 {
   if (depth > MAX_DEPTH)
   {
